+++ src/main/java/ru/javawebinar/topjava/model/Meal.java
@@ -6,25 +6,25 @@ import java.time.LocalTime;
 
 public class Meal extends AbstractBaseEntity {
 
-    private LocalDateTime dateTime;
+    private LocalDateTime mealDate;
     private String description;
     private int calories;
 
     public Meal() {}
 
-    public Meal(LocalDateTime dateTime, String description, int calories) {
-        this(null, dateTime, description, calories);
+    public Meal(LocalDateTime mealDate, String description, int calories) {
+        this(null, mealDate, description, calories);
     }
 
-    public Meal(Integer id, LocalDateTime dateTime, String description, int calories) {
+    public Meal(Integer id, LocalDateTime mealDate, String description, int calories) {
         super(id);
-        this.dateTime = dateTime;
+        this.mealDate = mealDate;
         this.description = description;
         this.calories = calories;
     }
 
-    public LocalDateTime getDateTime() {
-        return dateTime;
+    public LocalDateTime getMealDateTime() {
+        return mealDate;
     }
 
     public String getDescription() {
@@ -35,19 +35,19 @@ public class Meal extends AbstractBaseEntity {
         return calories;
     }
 
-    public LocalDate getDate() {
-        return dateTime.toLocalDate();
+    public LocalDate getMealDate() {
+        return mealDate.toLocalDate();
     }
 
-    public LocalTime getTime() {
-        return dateTime.toLocalTime();
+    public LocalTime getMealTime() {
+        return mealDate.toLocalTime();
     }
 
     @Override
     public String toString() {
         return "Meal{" +
                 "id=" + id +
-                ", dateTime=" + dateTime +
+                ", mealDate=" + mealDate +
                 ", description='" + description + '\'' +
                 ", calories=" + calories +
                 '}';