+++ src/main/java/ru/javawebinar/topjava/repository/jdbc/JdbcMealRepositoryImpl.java
@@ -5,7 +5,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.support.DataAccessUtils;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
-import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
@@ -14,16 +15,20 @@ import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 
 import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 
-@Repository
+@Repository("mealDao")
 public class JdbcMealRepositoryImpl implements MealRepository {
+
     private static final Logger LOGGER = LoggerFactory.getLogger(JdbcMealRepositoryImpl.class);
     private final static BeanPropertyRowMapper<Meal> ROW_MAPPER =
                                 BeanPropertyRowMapper.newInstance(Meal.class);
+
     private final SimpleJdbcInsert simpleJdbcInsert;
     private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
 
@@ -36,26 +41,24 @@ public class JdbcMealRepositoryImpl implements MealRepository {
                 .usingGeneratedKeyColumns("id");
     }
 
-
-
     @Override
     public Boolean save(Meal meal) {
-        LOGGER.info("Save meal: {}", meal);
-        MapSqlParameterSource parameterMap = new MapSqlParameterSource();
+//        LOGGER.info("Save meal: {}", meal);
+//        MapSqlParameterSource parameterMap = new MapSqlParameterSource();
+//
+//        parameterMap.addValue("user_id", AuthorizedUser.getId());
+//        parameterMap.addValue("description", meal.getDescription());
+//        parameterMap.addValue("calories", meal.getCalories());
+//        parameterMap.addValue("datetime", meal.getMealDate());
+//
+//        if (meal.isNew()) {
+//            Number mealId = simpleJdbcInsert.executeAndReturnKey(parameterMap);
+//            meal.setId(mealId.intValue());
+//        } else {
+//            simpleJdbcInsert.execute(parameterMap);
+//        }
 
-        parameterMap.addValue("user_id", AuthorizedUser.getId());
-        parameterMap.addValue("description", meal.getDescription());
-        parameterMap.addValue("calories", meal.getCalories());
-        parameterMap.addValue("datetime", meal.getDateTime());
-
-        if (meal.isNew()) {
-            Number mealId = simpleJdbcInsert.executeAndReturnKey(parameterMap);
-            meal.setId(mealId.intValue());
-        } else {
-            simpleJdbcInsert.execute(parameterMap);
-        }
-
-        return meal;
+        return null;
     }
 
     @Override
@@ -70,45 +73,57 @@ public class JdbcMealRepositoryImpl implements MealRepository {
 
     @Override
     public Meal getById(Integer id) {
-        String QUERY = "SELECT m.* FROM meals m WHERE m.id=? AND m.user_id=?";
-        List<Meal> meals = namedParameterJdbcTemplate
-                .getJdbcOperations()
-                .query(QUERY, ROW_MAPPER, id, AuthorizedUser.getId());
-
-        return DataAccessUtils.singleResult(meals);
+        Objects.requireNonNull(id);
+        String QUERY = "SELECT m.* FROM meals m WHERE m.id=:mealId AND m.user_id=:userId";
+        Map<String, Object> paramMap = new HashMap<>();
+        paramMap.put("mealId", id);
+        paramMap.put("userId", AuthorizedUser.getId());
+        return namedParameterJdbcTemplate.queryForObject(QUERY, paramMap, new MealMapper());
     }
 
     @Override
     public List<Meal> getAllByUserId(Integer userId) {
         Objects.requireNonNull(userId);
         LOGGER.info("Get all user id '{}' meals", userId);
-        String QUERY = "SELECT id, description, calories, datetime " +
-                        "FROM meals WHERE user_id=? ORDER BY datetime";
-        List<Meal> meals = namedParameterJdbcTemplate.getJdbcOperations()
-                .query(QUERY, (ResultSet resultSet, int i) -> {
-                            LOGGER.info("datetime = {}", resultSet.getString("datetime"));
-                            LocalDateTime dateTime =
-                                    LocalDateTime.parse(resultSet.getString("datetime"),
-                                            DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss"));
-                            Meal meal = new Meal(dateTime,
-                                resultSet.getObject("description", String.class),
-                                    resultSet.getInt("calories"));
-                                meal.setId(resultSet.getInt("id"));
-                                return meal;}
-                                , userId);
+        String QUERY = "SELECT id, description, calories, meal_date \n" +
+                        "FROM meals WHERE user_id=:userId ORDER BY meal_date";
+
+        Map<String, Object> paramMap = new HashMap<>();
+        paramMap.put("userId", AuthorizedUser.getId());
+        List<Meal> meals = namedParameterJdbcTemplate.query(QUERY, paramMap, new MealMapper());
         LOGGER.info("meals.size() = {}", meals.size());
         return meals;
     }
 
     @Override
     public List<Meal> getBetween(LocalDateTime startDate, LocalDateTime endDate, Integer userId) {
+        Objects.requireNonNull(startDate);
+        Objects.requireNonNull(endDate);
+        Objects.requireNonNull(userId);
 
-        LOGGER.info("getBetween()");
-        List<Meal> meals = namedParameterJdbcTemplate
-                .getJdbcOperations()
-                .query("SELECT id, description, datetime, calories " +
-                        "FROM meals " +
-                        "WHERE (datetime BETWEEN ? AND ?) AND user_id=?", ROW_MAPPER, startDate, endDate, userId);
-        return meals;
+        String QUERY = "SELECT id, description, meal_date, calories \n" +
+                        "FROM meals \n" +
+                        "WHERE (meal_date BETWEEN :startDate AND :endDate) AND user_id=:userId";
+
+        Map<String, Object> paramMap = new HashMap<>();
+        paramMap.put("startDate", startDate);
+        paramMap.put("endDate", endDate);
+        paramMap.put("userId", userId);
+        LOGGER.info("Get meal between {} and {} for user id {}", startDate, endDate, userId);
+        return namedParameterJdbcTemplate.query(QUERY, paramMap, new MealMapper());
+    }
+
+    private static final class MealMapper implements RowMapper<Meal> {
+        @Override
+        public Meal mapRow(ResultSet rs, int rowNum) throws SQLException {
+            Meal m = new Meal(
+                    rs.getInt("id"),
+                    rs.getTimestamp("meal_date").toLocalDateTime(),
+                    rs.getString("description"),
+                    rs.getInt("calories")
+            );
+
+            return m;
+        }
     }
 }