+++ src/main/java/ru/javawebinar/topjava/util/ValidationUtil.java
@@ -6,7 +6,7 @@ import ru.javawebinar.topjava.util.exception.NotFoundException;
 
 public class ValidationUtil {
 
-    public static void checkNotFoundWithId(boolean found, int id) {
+    public static void checkNotFoundWithId(Boolean found, int id) {
         checkNotFound(found, "id=" + id);
     }