+++ src/main/java/ru/javawebinar/topjava/web/meal/MealRestController.java
@@ -30,32 +30,32 @@ public class MealRestController {
     }
 
     public Meal get(int id) {
-        int userId = AuthorizedUser.id();
+        int userId = AuthorizedUser.getId();
         log.info("get meal {} for user {}", id, userId);
         return service.get(id, userId);
     }
 
     public void delete(int id) {
-        int userId = AuthorizedUser.id();
+        int userId = AuthorizedUser.getId();
         log.info("delete meal {} for user {}", id, userId);
         service.delete(id, userId);
     }
 
     public List<MealWithExceed> getAll() {
-        int userId = AuthorizedUser.id();
+        int userId = AuthorizedUser.getId();
         log.info("getAll for user {}", userId);
         return MealsUtil.getWithExceeded(service.getAll(userId), AuthorizedUser.getCaloriesPerDay());
     }
 
     public Meal create(Meal meal) {
-        int userId = AuthorizedUser.id();
+        int userId = AuthorizedUser.getId();
         checkNew(meal);
         log.info("create {} for user {}", meal, userId);
         return service.create(meal, userId);
     }
 
     public void update(Meal meal, int id) {
-        int userId = AuthorizedUser.id();
+        int userId = AuthorizedUser.getId();
         assureIdConsistent(meal, id);
         log.info("update {} for user {}", meal, userId);
         service.update(meal, userId);
@@ -68,7 +68,7 @@ public class MealRestController {
      * </ol>
      */
     public List<MealWithExceed> getBetween(LocalDate startDate, LocalTime startTime, LocalDate endDate, LocalTime endTime) {
-        int userId = AuthorizedUser.id();
+        int userId = AuthorizedUser.getId();
         log.info("getBetween dates({} - {}) time({} - {}) for user {}", startDate, endDate, startTime, endTime, userId);
 
         List<Meal> mealsDateFiltered = service.getBetweenDates(