+++ src/test/java/ru/javawebinar/topjava/repository/mock/JdbcUserRepositoryImplTest.java
@@ -1,0 +1,5 @@
+package ru.javawebinar.topjava.repository.mock;
+
+public class JdbcUserRepositoryImplTest {
+
+}